refactor: fetchAssetRates
